requirements.txt
opencv-python==4.5.4.60
numpy==1.19.5
matplotlib==3.3.4
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1